Add files via upload
Variational Autoencoders (VAEs) have emerged as powerful generative models that learn to
represent complex high-dimensional data in a low-dimensional latent space. They can generate
new data samples that are similar to the training data. In this report, we present the
implementation of a Variational Autoencoder for the MNIST digits dataset.
Variational Autoencoders (VAEs) are a type of generative model that learns to reconstruct input
data while simultaneously learning a latent representation of the data. By sampling from the
learned latent space, VAEs can generate new data that closely resembles the training data